linkedin/requirements.txt
attrs==25.3.0
beautifulsoup4==4.13.3
behave==1.2.6
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
colorama==0.4.6
cssselect==1.3.0
exceptiongroup==1.2.2
execnet==2.1.1
fasteners==0.19
filelock==3.18.0
h11==0.14.0
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mycdp==1.1.1
outcome==1.3.0.post0
packaging==24.2
parameterized==0.9.0
parse==1.20.2
parse_type==0.6.4
pdbp==1.7.0
platformdirs==4.3.7
pluggy==1.5.0
psycopg2-binary==2.9.10
Pygments==2.19.1
pynose==1.5.4
pyotp==2.9.0
PySocks==1.7.1
pytest==8.3.5
pytest-html==4.0.2
pytest-metadata==3.1.1
pytest-ordering==0.6
pytest-rerunfailures==15.0
pytest-xdist==3.6.1
PyYAML==6.0.2
requests==2.32.3
rich==14.0.0
sbvirtualdisplay==1.4.0
selenium==4.30.0
seleniumbase==4.36.5
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
tabcompleter==1.4.0
tomli==2.2.1
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.13.1
urllib3==2.3.0
websocket-client==1.8.0
websockets==15.0.1
wsproto==1.2.0
